+++ rust/operator-binary/src/backend/dynamic.rs
@@ -109,7 +109,7 @@ pub async fn from_class(
         crd::SecretClassBackend::KerberosKeytab(crd::KerberosKeytabBackend {
             realm_name,
             kdc,
-            admin_server,
+            admin,
             admin_keytab_secret,
             admin_principal,
         }) => from(
@@ -118,7 +118,9 @@ pub async fn from_class(
                 KerberosProfile {
                     realm_name,
                     kdc,
-                    admin_server,
+                    kadmin_server: match admin {
+                        crd::KerberosKeytabBackendAdmin::Mit { kadmin_server } => kadmin_server,
+                    },
                 },
                 &admin_keytab_secret,
                 admin_principal,